Pull out descriptors into their own files
This should make it clearer where each descriptor class lives and
should make it more straightforward to add new ones in the future.
