+++ pyremap/__init__.py
@@ -1,6 +1,9 @@
-from pyremap.descriptor import MpasMeshDescriptor, LatLonGridDescriptor, \
-    LatLon2DGridDescriptor, ProjectionGridDescriptor, \
-    PointCollectionDescriptor, get_lat_lon_descriptor
+from pyremap.mpas_mesh_descriptor import MpasMeshDescriptor
+from pyremap.lat_lot_grid_descriptor import LatLonGridDescriptor, \
+    get_lat_lon_descriptor
+from pyremap.lat_lot_2d_grid_descriptor import LatLon2DGridDescriptor
+from pyremap.projection_grid_descriptor import ProjectionGridDescriptor
+from pyremap.point_collection_descriptor import PointCollectionDescriptor
 
 from pyremap.remapper import Remapper
